+++ src/test/java/info/movito/themoviedbapi/TmdbChangesTest.java
@@ -3,17 +3,17 @@ package info.movito.themoviedbapi;
 import java.io.IOException;
 
 import info.movito.themoviedbapi.model.changes.ChangesResultsPage;
+import info.movito.themoviedbapi.testutil.TestUtils;
 import info.movito.themoviedbapi.tools.RequestType;
 import info.movito.themoviedbapi.tools.TmdbException;
-import info.movito.themoviedbapi.util.TestUtils;
 import org.junit.jupiter.api.Test;
 
 import static info.movito.themoviedbapi.TmdbChanges.TMDB_METHOD_CHANGES;
 import static info.movito.themoviedbapi.TmdbChanges.TMDB_METHOD_MOVIE;
 import static info.movito.themoviedbapi.TmdbChanges.TMDB_METHOD_PERSON;
 import static info.movito.themoviedbapi.TmdbChanges.TMDB_METHOD_TV;
+import static info.movito.themoviedbapi.testutil.TestUtils.validateAbstractJsonMappingFields;
 import static info.movito.themoviedbapi.tools.ApiUrl.TMDB_API_BASE_URL;
-import static info.movito.themoviedbapi.util.TestUtils.validateAbstractJsonMappingFields;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.Mockito.when;
