+++ src/test/java/info/movito/themoviedbapi/TmdbCompaniesTest.java
@@ -5,14 +5,14 @@ import java.io.IOException;
 import info.movito.themoviedbapi.model.companies.AlternativeNamesResultsPage;
 import info.movito.themoviedbapi.model.companies.Company;
 import info.movito.themoviedbapi.model.core.image.ImageResults;
+import info.movito.themoviedbapi.testutil.TestUtils;
 import info.movito.themoviedbapi.tools.RequestType;
 import info.movito.themoviedbapi.tools.TmdbException;
-import info.movito.themoviedbapi.util.TestUtils;
 import org.junit.jupiter.api.Test;
 
 import static info.movito.themoviedbapi.TmdbCompanies.TMDB_METHOD_COMPANY;
+import static info.movito.themoviedbapi.testutil.TestUtils.validateAbstractJsonMappingFields;
 import static info.movito.themoviedbapi.tools.ApiUrl.TMDB_API_BASE_URL;
-import static info.movito.themoviedbapi.util.TestUtils.validateAbstractJsonMappingFields;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.Mockito.when;
